chore: cleanup config
