+++ packages/fuel-indexer-lib/Cargo.toml
@@ -9,6 +9,7 @@ description = "Fuel Indexer Library"
 [dependencies]
 anyhow = "1.0"
 clap = { version = "3.1", features = ["cargo", "derive", "env"] }
+http = { version = "0.2", default-features = false }
 serde = { version = "1.0", features = ["derive"] }
 serde_yaml = "0.8"
 sha2 = "0.9"
